Fix upstart script at Advanced Deployment section
It includes a missing "=" in order to make the script valid.
The script was validated with the `init-checkconf` command.

This also includes a small change in the name of upstart script for the
example application. Upstart should lookup the script by its name, so
the command `start hello_phoenix` will use the
`/etc/init/hello_phoenix.conf` file.