Create A_custom_primary_key.md
My motivation for this document: it is important to be able to create a primary key other than id:integer in a variety of situations, especially if Phoenix is being adopted in existing projects or organizations.  I wanted to have all the steps in one place and easily accessible to Phoenix users -- especially new Phoenix users like myself.   Big thanks to Jose Valim and Slack users  @sg, @angeloashmore, @gjaldon for their help and feedback!