Postgres Installation: Note about default encoding
On OSX, when Postgres is installed using brew, the default encoding is set to SQL ASCII as explained https://www.postgresql.org/docs/current/static/app-initdb.html. This caused an error when doing `mix ecto.create` in the Up and Running section. Creating Postgres DB with the UTF-8 fixes this problem. 

I believe this "note" is needed because people who are new to Postgres and starting up with Elixir and Phoenix might think this is a Phoneix/Elixir/Erlang issue and not necessarily Postgres issue.