Recommend an easier intall of Postgres for Mac users
I wanted to add an easier install of Postgres for Mac users with the Postgres App, a Postgres Mac GUI. This seemed to be the easiest way to get up and running with Postgres on my computer. I saw in other forums people were having dificulty with this as well, that's how I found Postgres App. Hopefully this can help others too.

I also fixed a type on line 92 where is said Postgrex.